% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check.Nexus.Names.R
\name{Check.Nexus.Names}
\alias{Check.Nexus.Names}
\title{Check.Nexus.Names: Function to check the overlap in taxnames between a vector (provided from a file usually) and a nexus file}
\usage{
Check.Nexus.Names(vector.TaxaNames, handle.NexusFile)
}
\arguments{
\item{vector.TaxaNames}{Vector containing the taxa names to be evaluated}

\item{handle.NexusFile}{Handle object of nexus file obtained from (read.nexus.data from APE)}
}
\description{
This function returns a list containing (1) a vector of names found in nexus file and (2) a vector of names that are not found in the nexus file
}
\examples{
library(ape)
library(RepeatLandscapeR)

vector.TaxaNames <- as.vector(read.table(file = '~/Copy/AmphibianGenomics/Methods&Materials/Step1_PhylogeneticInference/Step1.4_ReNameNexus/TaxaNameMapSheet.txt', header = T)[,2])
handle.NexusFile <- read.nexus.data(file = '~/Copy//AmphibianGenomics/Methods&Materials/Step1_PhylogeneticInference/Step1.4_ReNameNexus/Pyron2011_Amphib.nexus')

CheckName.Results <- Check.Nexus.Names(vector.TaxaNames = vector.TaxaNames, handle.NexusFile = handle.NexusFile)
}
\keyword{comparative}
\keyword{elements,}
\keyword{evolution}
\keyword{genomics,}
\keyword{repeat}
