% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check.Tree.Names.R
\name{Check.Tree.Names}
\alias{Check.Tree.Names}
\title{Check.Tree.Names: Function to check the overlap in taxnames between a vector (provided from a file usually) and a newick tree}
\usage{
Check.Tree.Names(vector.TaxaNames, handle.Tree)
}
\arguments{
\item{vector.TaxaNames}{Vector containing the taxa names to be evaluated}

\item{handle.Tree}{Handle object of newick tree obtained from (read.tree from APE)}
}
\description{
This function returns a list containing (1) a vector of names found in newick tree and (2) a vector of names that are not found in the newick tree
}
\examples{
library(ape)
library(RepeatLandscapeR)

}
\keyword{comparative}
\keyword{elements,}
\keyword{evolution}
\keyword{genomics,}
\keyword{repeat}
