% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.Nexus.R
\name{Filter.Nexus}
\alias{Filter.Nexus}
\title{Filter.Nexus: Function to filter a large nexus file to only contain sequences from a given vector of names}
\usage{
Filter.Nexus(vector.TaxaNames, handle.NexusFile, path.OutNexus)
}
\arguments{
\item{vector.TaxaNames}{Vector containing the taxa names to be evaluated}

\item{handle.NexusFile}{Handle object of nexus file obtained from (read.nexus.data from APE)}
}
\description{
This function writes a new nexus file with only sequences from taxa contained in the vectort.TaxaNames
}
\examples{
library(ape)
library(RepeatLandscapeR)

vector.TaxaNames <- as.vector(read.table(file = '~/Copy/AmphibianGenomics/Methods&Materials/Step1_PhylogeneticInference/Step1.5_FilterNexus/TaxaNameMapSheet.txt', header = T)[,2])
handle.NexusFile <- read.nexus.data(file = '~/Copy//AmphibianGenomics/Methods&Materials/Step1_PhylogeneticInference/Step1.5_FilterNexus/Pyron2011_Amphib.nexus')
Filter.Nexus(vector.TaxaNames = vector.TaxaNames, handle.NexusFile = handle.NexusFile, path.OutNexus = '~/Copy/AmphibianGenomics/Methods&Materials/Step1_PhylogeneticInference/Step1.5_FilterNexus/Filtered.nexus')
}
\keyword{comparative}
\keyword{elements,}
\keyword{evolution}
\keyword{genomics,}
\keyword{repeat}
