% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_ConcateProcessedReads.R
\name{Pipeline_ConcateProcessedReads}
\alias{Pipeline_ConcateProcessedReads}
\title{Pipeline_ConcateProcessedReads.R: Function to X}
\usage{
Pipeline_ConcateProcessedReads(dir.Parent, path.GenomeFastqDir)
}
\arguments{
\item{X}{X}
}
\description{
This function returns XXX
}
\examples{
library(ape)
library(RepeatLandscapeR)

Pipeline_ConcateProcessedReads(dir.Parent = '~/Copy/AmphibianGenomics/Methods&Materials/Step0_ProcessReads/Step0.2_ConcateProcessedReads_PIPELINE/', 
  path.GenomeFastqDir = '~/Copy/AmphibianGenomics/Methods&Materials/Step0_ProcessReads/Step0.2_ConcateProcessedReads_PIPELINE/GenomeFastqFiles/')
}
\keyword{comparative}
\keyword{elements,}
\keyword{evolution}
\keyword{genomics,}
\keyword{repeat}
