% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline_PALFinder.R
\name{Pipeline_PALFinder}
\alias{Pipeline_PALFinder}
\title{Pipeline_PALFinder.R: Function to conduct PAL finder analyses on a directory of genome fasta files}
\usage{
Pipeline_PALFinder(dir.Parent, path.GenomeFastaDir, path.Primer3,
  path.PAL_finder)
}
\arguments{
\item{dir.Parent}{Path to parent directory, subdirectory for all analyses will be created here}

\item{path.GenomeFastaDir}{Path to directory containing the fasta genome files}

\item{path.Primer3}{Path to Primer3_core executable (used by Palfinder)}

\item{path.PAL_finder}{Path to pal_finder executable}
}
\description{
This function returns a directory containg the results of a PAL finder analysis on a set of genome fasta files
}
\examples{
library(ape)
library(RepeatLandscapeR)

Pipeline_PALFinder(dir.Parent = '~/Copy/AmphibianGenomics/Methods&Materials/Step2_PAL_Finder/Step0_Examples/PALFinderAnalysis/', 
  path.GenomeFastaDir = '~/Copy/AmphibianGenomics/Methods&Materials/Step2_PAL_Finder/Step0_Examples/PALFinderAnalysis/Genomes_FastaFiles/', 
  path.Primer3 = '/Applications/primer3-2.4.0/src/primer3_core', 
  path.PAL_finder = '/Applications/pal_finder_v0.02.04.pl')
}
\keyword{comparative}
\keyword{elements,}
\keyword{evolution}
\keyword{genomics,}
\keyword{repeat}
