% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write.PALfinder.ConfigFile.R
\name{Write.PALfinder.ConfigFile}
\alias{Write.PALfinder.ConfigFile}
\title{Write.PALfinder.ConfigFile: Function to write a configuration file for PAL finder}
\usage{
Write.PALfinder.ConfigFile(path.ConfigFile, numeric.FindPrimers,
  string.Platform, string.InputFormat, numeric.PE, path.454readfile,
  path.inputreadfile, path.inputPEreadfile, path.MicrosateSumout,
  path.PALsummaryOut, path.Primer3)
}
\arguments{
\item{path.ConfigFile}{Path to output configuration file}

\item{numeric.FindPrimers}{Numeric answer (0 = no, 1 = yes)}

\item{string.Platform}{String for sequencing platform ("Illumina" or "454")}

\item{string.InputFormat}{String for input format ("fasta" or "fastq")}

\item{numeric.PE}{Path to fasta genome file}

\item{path.inputreadfile}{Path to fastq file read1 (Illumina)}

\item{path.inputPEreadfile}{Path to fastq file read2 (Illumina)}

\item{path.MicrosateSumout}{Path to output microsatellite summary statistics}

\item{path.PALsummaryOut}{Path to output PAL summary statistics}

\item{path.Primer3}{Path to Primer3_core executable}
}
\description{
This function returns writes a series of information into the PALfinder configuration file set by the user
}
\examples{
library(ape)
library(RepeatLandscapeR)


Write.PALfinder.ConfigFile(path.ConfigFile = '~/Copy/AmphibianGenomics/Methods&Materials/Step2_PAL_Finder/Step0_Examples/PALFinderAnalysis/config.txt', 
numeric.FindPrimers = 0, 
string.Platform = "454", 
string.InputFormat = "fasta", 
numeric.PE = 0, 
path.454readfile = 'SAMPLED_RMB2503_Rhyacotriton-kezeri.fasta', 
path.MicrosateSumout = 'test_microsat_summary.txt', 
path.PALsummaryOut = 'test_PAL_summary.txt', 
path.Primer3 = '/Applications/primer3-2.4.0/src/primer3_core')
}
\keyword{comparative}
\keyword{elements,}
\keyword{evolution}
\keyword{genomics,}
\keyword{repeat}
